Change ScannerStateService#cleansing to package private
 - This field is mostly a test flag, currently.
 - Integrate test cases and fix fields to not expose
